move mixer_dev.py 'snps' feature from gsa_mixer to mixer_dev
