+++ precimed/mixer_dev/cli.py
@@ -11,6 +11,7 @@ import scipy.optimize
 import scipy.io as sio
 import time
 import sys
+import random
 
 from datetime import datetime
 from pathlib import Path
@@ -166,6 +167,18 @@ def parser_add_arguments(parser, func, analysis_type):
             parser.add_argument('--ci-alpha', type=float, default=None, help=argparse.SUPPRESS)              # significance level for the confidence interval estimation
             parser.add_argument('--ci-samples', type=int, default=10000, help=argparse.SUPPRESS)             # number of samples in uncertainty estimation
             parser.add_argument('--ci-power-samples', type=int, default=100, help=argparse.SUPPRESS)         # number of samples in power curves uncertainty estimation
+
+    parser.set_defaults(func=func)
+
+def parser_snps_add_arguments(func, parser):
+    parser.add_argument('--r2', type=float, default=None, help="LD r2 threshold for prunning SNPs (default: %(default)s)")
+    parser.add_argument('--maf', type=float, default=None, help="minor allele frequence (MAF) threshold (default: %(default)s)")
+    parser.add_argument('--subset', type=int, default=2000000, help="number of SNPs to randomly select (default: %(default)s)")
+    parser.add_argument('--seed', type=np.uint32, default=None, help="random seed (default: %(default)s)")
+
+    common.utils_cli.parser_add_argument_bim_file(parser)
+    common.utils_cli.parser_add_argument_ld_file(parser)
+    common.utils_cli.parser_add_argument_chr2use(parser)
 
     parser.set_defaults(func=func)
 
@@ -822,4 +835,45 @@ def execute_fixed_effects_parser(args):
     bim_tag['BETA'] = fixedeffectbeta
     bim_tag['PVAL'] = 1
     bim_tag.to_csv(args.out + '.fixed_effects.csv', sep='\t', index=False)
+    libbgmg.log_message('Done')
+
+
+def execute_snps_parser(args):
+    common.utils_cli.fix_and_validate_args(args)
+
+    libbgmg = gsa_mixer.cli.initialize_mixer_plugin(args, 0, "", "", None, None, None, args.chr2use, '@')
+    mafvec = np.minimum(libbgmg.mafvec, 1-libbgmg.mafvec)
+
+    libbgmg.log_message('Load {}...'.format(args.bim_file))
+    ref=pd.concat([pd.read_csv(args.bim_file.replace('@', str(chr_label)), sep='\t', header=None, names='CHR SNP GP BP A1 A2'.split()) for chr_label in args.chr2use])
+    libbgmg.log_message('{} SNPs in total'.format(len(ref)))
+
+    # step0 - generate random values for clumping (this is a way to implement random pruning)
+    buf = np.random.rand(libbgmg.num_tag,)
+    
+    # step1 - filter SNPs below MAF threshold
+    if args.maf:
+        buf[mafvec<args.maf] = np.nan
+        libbgmg.log_message('{} SNPs pass --maf {} threshold'.format(np.sum(np.isfinite(buf)), args.maf))
+    
+    # step2 - select a random subset of SNPs
+    if args.subset > 0:
+        indices = list(np.where(np.isfinite(buf))[0])
+        sample = random.sample(indices, min(args.subset, len(indices)))
+        mask = np.ones(buf.shape); mask[sample] = 0
+        buf[mask == 1] = np.nan
+        libbgmg.log_message('{} SNPs randomly selected (--subset)'.format(np.sum(np.isfinite(buf))))
+
+    # step3 - further prune SNPs at certain r2 threshold
+    if args.r2:
+        buf = libbgmg.perform_ld_clump(args.r2, buf.flatten()) 
+        libbgmg.log_message('{} SNPs pass random prunning at --r2 {} threshold'.format(np.sum(np.isfinite(buf)), args.r2))
+
+    # step4 - save results
+    ref.SNP[np.isfinite(buf)].to_csv(args.out, index=False, header=False)
+    libbgmg.log_message('Result saved to {}'.format(args.out))
+
+    # step 5 - save .info file
+    #pd.DataFrame({'chrnumvec': libbgmg.chrnumvec, 'mafvec': libbgmg.mafvec, 'tldvec': libbgmg.ld_sum_r2}).to_csv(args.out + '.info', sep='\t', index=False)
+    
     libbgmg.log_message('Done')
